Bouton 'supprimer' dans 'voir un contrat'; bientôt terminé, j'espère.

+++ src/Controller/FrontEnd/CompteClientController.php
@@ -40,10 +40,12 @@ final class CompteClientController extends AbstractController {
     #[Route('/listeDesContrats/{id}' , name: 'app_liste_des_contrats', methods: ['GET'])]
     public function listeDesContrats(Request $request, ContratRepository $contratRepository, UtilisateurRepository $utilisateurRepository) : Response {
 
-        $pathContratActuelDansPublic = str_replace( "//" , "/" , $this->params->get('app.public_upload_dir') . $request->query->get( 'pathContratDansPublic' ) );
-        $pathContratActuelDansPublic = str_replace( "/storage/" , "/" , $pathContratActuelDansPublic );
-        if ( file_exists( $pathContratActuelDansPublic ) ) {
-            unlink( $pathContratActuelDansPublic );
+        if ( ! empty( $request->query->get( 'pathContratDansPublic' ) ) ) {
+            $pathContratActuelDansPublic = str_replace( "//" , "/" , $this->params->get('app.public_upload_dir') . $request->query->get( 'pathContratDansPublic' ) );
+            $pathContratActuelDansPublic = str_replace( "/storage/" , "/" , $pathContratActuelDansPublic );
+            if ( file_exists( $pathContratActuelDansPublic ) ) {
+                unlink( $pathContratActuelDansPublic );
+            }
         }
 
         $id = $request->attributes->get( 'id' );
